Finish the environ mock up and the storage volume unit tests
